sheet_init.py
import gspread
from google.oauth2.service_account import Credentials
import os
from dotenv import load_dotenv

load_dotenv()

def initialize_sheets():
    # Placeholder for sheet initialization logic
    # Requires credentials.json from Google Cloud Console
    print("Sheet initialization script template created.")
    print("Required columns: Brand, Source Type, Identifier, Last Checked, Last Processed ID, Status, Blog Ref, Social Ref, Timestamp, Error Logs")

if __name__ == "__main__":
    initialize_sheets()
